pyproject.toml
[project]
name = "puzzle-bot"
version = "0.1.0"
description = "Discord puzzle bot with sqlite state"
readme = "README.md"
authors = [
    { name = "ioni", email = "104294646+IongIer@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
